make storage lib separate so we can build without storage libs
