+++ .gitmodules
@@ -1,5 +1,5 @@
-[submodule "ios/mapbox-gl-cocoa"]
-	path = ios/mapbox-gl-cocoa
+[submodule "ios"]
+	path = ios
 	url = https://github.com/mapbox/mapbox-gl-cocoa.git
 
 [submodule "test/suite"]
@@ -13,3 +13,6 @@
 [submodule "styles"]
 	path = styles
 	url = https://github.com/mapbox/mapbox-gl-styles.git
+[submodule "app/ios"]
+	path = app/ios
+	url = https://github.com/mapbox/mapbox-gl-cocoa.git
