+++ bin/render.gyp
@@ -44,10 +44,10 @@
       ],
       'include_dirs': [ '../src' ],
       'dependencies': [
-        '../mapboxgl.gyp:mbgl-standalone',
-        '../mapboxgl.gyp:mbgl-headless',
-        '../mapboxgl.gyp:mbgl-<(platform)',
-        '../mapboxgl.gyp:copy_certificate_bundle',
+        '../mbgl.gyp:<(core_library)',
+        '../mbgl.gyp:<(headless_library)',
+        '../mbgl.gyp:<(platform_library)',
+        '../mbgl.gyp:copy_certificate_bundle',
       ],
     },
   ],
