+++ gyp/mbgl-ios.gypi
@@ -8,34 +8,47 @@
       'dependencies': [
         'version',
       ],
+
       'sources': [
         '../platform/darwin/log_nslog.mm',
         '../platform/darwin/string_nsstring.mm',
-        '../platform/darwin/http_request_cocoa.mm',
         '../platform/darwin/application_root.mm',
         '../platform/darwin/image.mm',
-        '../platform/default/asset_request_libuv.cpp',
       ],
-      'include_dirs': [
-        '../include',
-        '../src',
-      ],
-      'xcode_settings': {
-        'OTHER_CPLUSPLUSFLAGS': [
+
+      'variables': {
+        'cflags_cc': [
           '<@(uv_cflags)',
           '-I<(boost_root)/include',
         ],
+        'libraries': [
+          '<@(uv_static_libs)',
+        ],
+        'ldflags': [
+          '-framework ImageIO',
+          '-framework MobileCoreServices',
+        ],
       },
+
+      'include_dirs': [
+        '../include',
+      ],
+
+      'xcode_settings': {
+        'OTHER_CPLUSPLUSFLAGS': [ '<@(cflags_cc)' ],
+      },
+
+      'link_settings': {
+        'libraries': [ '<@(libraries)' ],
+        'xcode_settings': {
+          'OTHER_LDFLAGS': [ '<@(ldflags)' ],
+        },
+      },
+
       'direct_dependent_settings': {
         'include_dirs': [
           '../include',
         ],
-        'xcode_settings': {
-          'OTHER_LDFLAGS': [
-            '-framework ImageIO',
-            '-framework MobileCoreServices',
-          ],
-        },
       },
     },
   ],
