+++ gyp/mbgl-osx.gypi
@@ -8,35 +8,49 @@
       'dependencies': [
         'version',
       ],
+
       'sources': [
         '../platform/darwin/log_nslog.mm',
         '../platform/darwin/string_nsstring.mm',
-        '../platform/darwin/http_request_cocoa.mm',
         '../platform/darwin/application_root.mm',
         '../platform/darwin/image.mm',
-        '../platform/default/asset_request_libuv.cpp',
       ],
+
+      'variables': {
+        'cflags_cc': [
+          '<@(uv_cflags)',
+          '-I<(boost_root)/include',
+        ],
+        'libraries': [
+          '<@(uv_static_libs)',
+        ],
+        'ldflags': [
+          '-framework Foundation',
+          '-framework ImageIO',
+          '-framework CoreServices',
+        ],
+      },
+
       'include_dirs': [
         '../include',
         '../src',
       ],
+
       'xcode_settings': {
-        'OTHER_CPLUSPLUSFLAGS': [
-          '<@(uv_cflags)',
-          '-I<(boost_root)/include',
-        ],
+        'OTHER_CPLUSPLUSFLAGS': [ '<@(cflags_cc)' ],
       },
+
+      'link_settings': {
+        'libraries': [ '<@(libraries)' ],
+        'xcode_settings': {
+          'OTHER_LDFLAGS': [ '<@(ldflags)' ],
+        },
+      },
+
       'direct_dependent_settings': {
         'include_dirs': [
           '../include',
         ],
-        'xcode_settings': {
-          'OTHER_LDFLAGS': [
-            '-framework Foundation',
-            '-framework ImageIO',
-            '-framework CoreServices',
-          ],
-        },
       },
     },
   ],
