--- gyp/mbgl-platform.gypi
@@ -1,20 +1,0 @@
-{
-  'conditions': [
-    ['platform == "osx"', {
-      'includes': ['mbgl-osx.gypi'],
-      'variables': { 'platform_library': 'mbgl-osx' },
-    }],
-    ['platform == "ios"', {
-      'includes': ['mbgl-ios.gypi'],
-      'variables': { 'platform_library': 'mbgl-ios' },
-    }],
-    ['platform == "linux"', {
-      'includes': ['mbgl-linux.gypi'],
-      'variables': { 'platform_library': 'mbgl-linux' },
-    }],
-    ['platform == "android"', {
-      'includes': ['mbgl-android.gypi'],
-      'variables': { 'platform_library': 'mbgl-android' },
-    }],
-  ],
-}
