+++ gyp/mbgl-storage-curl.gypi
@@ -1,62 +1,52 @@
 {
   'targets': [
-    { 'target_name': 'mbgl-linux',
-      'product_name': 'mbgl-linux',
+    { 'target_name': 'mbgl-storage-curl',
+      'product_name': 'mbgl-storage-curl',
       'type': 'static_library',
       'standalone_static_library': 1,
       'hard_dependency': 1,
-      'dependencies': [
-        'version',
-      ],
 
       'sources': [
-        '../platform/default/log_stderr.cpp',
-        '../platform/default/string_stdlib.cpp',
+        '../platform/default/default_file_source.cpp',
+        '../platform/default/sqlite_cache.cpp',
         '../platform/default/http_request_curl.cpp',
         '../platform/default/asset_request_libuv.cpp',
-        '../platform/default/application_root.cpp',
-        '../platform/default/image.cpp',
-        '../platform/default/image_reader.cpp',
-        '../platform/default/png_reader.cpp',
-        '../platform/default/jpeg_reader.cpp',
+        '../platform/default/sqlite3.hpp',
+        '../platform/default/sqlite3.cpp',
+        '../platform/default/compression.hpp',
+        '../platform/default/compression.cpp',
+      ],
+
+      'include_dirs': [
+        '../include',
       ],
 
       'variables': {
         'cflags_cc': [
-          '<@(png_cflags)',
-          '<@(jpeg_cflags)',
           '<@(uv_cflags)',
-          '<@(nu_cflags)',
+          '<@(sqlite3_cflags)',
           '-I<(boost_root)/include',
         ],
         'ldflags': [
-          '<@(png_ldflags)',
-          '<@(jpeg_ldflags)',
           '<@(uv_ldflags)',
-          '<@(curl_ldflags)',
-          '<@(nu_ldflags)',
+          '<@(sqlite3_ldflags)',
+          '<@(zlib_ldflags)',
         ],
         'libraries': [
-          '<@(png_static_libs)',
-          '<@(jpeg_static_libs)',
           '<@(uv_static_libs)',
-          '<@(nu_static_libs)',
+          '<@(sqlite3_static_libs)',
+          '<@(zlib_static_libs)',
         ],
       },
-
-      'include_dirs': [
-        '../include',
-        '../src',
-      ],
 
       'conditions': [
         ['OS == "mac"', {
           'xcode_settings': {
             'OTHER_CPLUSPLUSFLAGS': [ '<@(cflags_cc)' ],
-          }
+          },
         }, {
-          'cflags_cc': [ '<@(cflags_cc)' ],
-        }]
+         'cflags_cc': [ '<@(cflags_cc)' ],
+        }],
       ],
 
       'link_settings': {
@@ -71,9 +61,7 @@
       },
 
       'direct_dependent_settings': {
-        'include_dirs': [
-          '../include',
-        ],
+        'include_dirs': [ '../include' ],
       },
     },
   ],
