--- mapboxgl.gyp
@@ -1,13 +1,0 @@
-{
-  'includes': [
-    './gyp/common.gypi',
-    './gyp/shaders.gypi',
-    './gyp/version.gypi',
-    './gyp/styles.gypi',
-    './gyp/certificates.gypi',
-    './gyp/mbgl-core.gypi',
-    './gyp/mbgl-platform.gypi',
-    './gyp/mbgl-headless.gypi',
-    './gyp/install.gypi',
-  ],
-}
