+++ platform/default/compression.hpp
