+++ platform/default/default_file_source.cpp
